added trace info that wraps up all the info that the delegate can provide and other sanitized names

+++ instrument/utils.go
@@ -21,24 +21,24 @@ import (
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-func computeApproximateRequestSize(r *http.Request) int {
-	s := 0
+func computeApproximateRequestSize(r *http.Request) int64 {
+	s := int64(0)
 	if r.URL != nil {
-		s += len(r.URL.String())
+		s += int64(len(r.URL.String()))
 	}
 
-	s += len(r.Method)
-	s += len(r.Proto)
+	s += int64(len(r.Method))
+	s += int64(len(r.Proto))
 	for name, values := range r.Header {
-		s += len(name)
+		s += int64(len(name))
 		for _, value := range values {
-			s += len(value)
+			s += int64(len(value))
 		}
 	}
-	s += len(r.Host)
+	s += int64(len(r.Host))
 
 	if r.ContentLength != -1 {
-		s += int(r.ContentLength)
+		s += r.ContentLength
 	}
 	return s
 }
